split scheduler into modular process & core_manager
